Update README.md
[unordered-list-marker-style] Marker style should be `*`
[no-heading-punctuation] Don’t add a trailing `:` to headings
Time to fix: 5 minutes

Why is this an issue?
Warn when a heading ends with a group of characters.

Options: string, default: '.,;:!?'.

Note: these are added to a regex, in a group ('[' + char + ']'), be careful to escape the string correctly.

Related code pattern
Warn when a heading ends with a group of characters.
Warn when a heading ends with a group of characters.

Markdown Code Style
Powered by RemarkLint