Update README.md
[no-undefined-references] Found reference to undefined definition
Time to fix: 5 minutes

Why is this an issue?
Warn when references to undefined definitions are found.

Options: Object, optional.

The object can have an allow property, an array of strings that may appear between [ and ] but that should not be treated as link identifiers.

Related code pattern
Warn when references to undefined definitions are found.
Warn when references to undefined definitions are found.

Markdown Code Style
Powered by RemarkLint