Update scrabble_score.py
scrabble_score_generator is responsible for returning a dictionary of letter values, so I think it would be more than ok for it to contain the lookup table of letter groups, which would shorten the solution to two functions:

scrabble_score_generator - which converts from the grouped lookup table, to a new dictionary with a score for every letter.
score() - which takes a word, and uses a dictionary of letter scores to return the word score.
Another thing to look at is only looping through the original table of letter groups once, instead of 26 times. dict() objects are unordered, so you don't actually need string.ascii_uppercase at all. You can loop through the "grouped" dictionary once, instead of repeatedly searching for scores. Try something like:

scrabble_score_table = dict()

for score in letter_value.keys():
    for letter in letter_value[score]:
        scrabble_score_table[letter]=score