Update kindergarten_garden.py
Splitting the diagram into its rows using integer division and slicing is clever, but you could use split() for this. I think it would be easier to understand. It's also simple enough that you don't need to assign it to a new variable, you could just use it directly.
The default students list is already in order, so good job on not sorting it again. I would move this into a module constant (e.g. an ALL_CAPS variable at the top of the file would be the common python style for constants. Similarly the _plants_converter could be made into a module constant.