Update kindergarten_garden.py
Other than those minor changes, I think the calculation in plants() could be more efficient and easier to read.

You are using an efficient slicing technique to get the plants from each row of the diagram, but you are doing it twice (once for each row). A nested for loop, or a nested list comprehension could do this in a single operation. Since self._diagram is a list with two lines from the diagram in it, you can loop over it directly instead of accessing each of it's indexes in separate for loops.