Update kindergarten_garden.py
Using while is good when you need to test an unknown condition, and don't know how many times you'll need to loop. If you know how many times you'll need to loop, a for loop with range is more predictable and easier to understand. It also also you won't need to keep track of a counter variable (row in this case). However, in many cases you want to loop over the data itself, rather than some number of times. If what you want, as in this case, is to do something for each item in an iterable (like your _diagram list here), you can just use:for row in self._diagram and each loop will load one item from the list into row.