Update grade_school.py
sorted(iter) returns a new, sorted list. Creating a new list requires copying all of the elements from the old one. Instead of creating a new list, it is faster to sort the existing list in-place using list.sort(). Keep in mind that anyone else who has a reference to the same list will also see that list become sorted, so be careful not to modify a list that something else is expecting to remain constant. In this case, the list is "owned" by you: you have modified it in the previous line and you (hopefully) won't be giving others that same list, so that they can't modify your list (we will cover this shortly).

If you find yourself looking to see if something is in a dict, and if not, giving it a default value, consider using dict.get(key, default), dict.setdefault(key, default), or the class collections.defaultdict.

If you find yourself building a new list by iterating over some other list and performing an operation on it, consider using a list comprehension. This is a compact expression that lets you define the new list, the for loop, and the operation to be performed on each item; all in a single, concise expression.

Also, if you want to append all the items from one list (source) on to another list (destination), rather than iterating over the source list (for item in source:) and appending each item to the destination (destination.append(item)), you can just use destination.extend(source). This will perform that copy for you. It will run much faster and provides a more concise description in your code.

Finally, note that grade() is returning one of your lists from within your dictionary. This means the caller now has a reference to that same list: if they modify the list, they are modifying the same list your are keeping internally, and now other callers will also see those changes. That's probably not what you want. Instead, you should make a copy of the list so that you are safe from any modifications the caller might make. There are several ways to do this. You could use list(existing) to make a new list based on the existing list; existing[:] to take a slice of the entire list, which also results in a new list; or existing.copy() which uses a method specific to the list type to create a copy very efficiently.