New method implementation
Rather than having the Team decode this, what if we just told it whether it had a win, loss, or draw? Then it wouldn't matter how we came to know this, we would just have to inform the team. The less that the Team knows about how we determine the match results, the less likely it will be that the Team will need to be modified if we modify how we represent match results.