Update rational_numbers.py
For __truediv__(), you don't really need to test and raise error, because now self.denom can't be null, and if other.numer is, it's already a ZeroDivisionError.
For __pow__(), there is no difference between int or float power (you return the same thing), so the only case is about the sign of power. Try to isolate the only thing that it changes about what you have to return.